+++ CHANGELOG.md
@@ -12,6 +12,7 @@ changes.
 
 ### Added
 - Black Consciousness Day ('Dia Nacional de Zumbi e da Consciência Negra') is public holiday in Brazil. [\#365](https://github.com/azuyalabs/yasumi/pull/365) ([c960657](https://github.com/c960657))
+- Christmas Eve is a public holiday in Poland from 2025. [\#370](https://github.com/azuyalabs/yasumi/pull/370) ([c960657](https://github.com/c960657))
 
 ### Changed